pyproject.toml
[project]
name = "discord-linker-frontend"
version = "0.1.0"
description = ""

requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "python-dotenv",
    "asyncpg",
    "websockets"
]

[tool.poetry]
packages = [{include = "discord_linker_frontend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
discord-linker-frontend = "discord_linker_frontend.__main__:main"
